+++ oc-includes/osclass/default-constants.php
@@ -13,7 +13,7 @@
  */
 
 if (!defined('OSCLASS_VERSION')) {
-    define('OSCLASS_VERSION', '6.0.1');
+    define('OSCLASS_VERSION', '6.0.2');
 }
 
 if (!defined('MULTISITE')) {